w00t

+++ src-tauri/migrations/2024-09-30-215016_nullable_artist/up.sql
@@ -16,7 +16,7 @@ CREATE TABLE beats (
     cover_art VARCHAR,
     comments TEXT,
     file_path VARCHAR NOT NULL,
-    bpm INTEGER,
+    bpm DOUBLE,
     musical_key VARCHAR,
     date_created DATETIME NOT NULL DEFAULT current_timestamp
 );
